requirements.txt
bs4
pymongo
pyyaml
redis